Fix rendering of ordered lists
The type of the list is passed by the flags parameter.  So if we
encounter an ORDERED list we use the list_counter to format a numbered
list.  list() gets called whenever a new list starts, so we can reset
the list_counter there.

This probably doesn't work with nested lists.

Fixes #5